pyproject.toml
[tool.poetry]
name = "browser-use-example"
version = "0.1.0"
description = ""
authors = ["HirokiYasui <hiromaily2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.49.1"
browser-use = "^0.1.17"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"
pytest = "^8.3.4"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
